refactor: improved PetProfileIndex
